Implemented get_wiget, config
get_wiget: took a example from google for grabbing a object from s3 but added the select one clause and turning it into a class

config: and gpt build out the argparse from the specific required arguments to add. pretty over engeneerd but it was simple just tedious now I have everthing I need to
continue with the rest of the program.